Conflicts:
	src/InfoBoxes/InfoBoxSettings.cpp

+++ src/Form/DataField/Integer.hpp
@@ -52,7 +52,7 @@ public:
   DataFieldInteger(const TCHAR *edit_format, const TCHAR *display_format,
                    int _min, int _max, int _value, int _step,
                    DataFieldListener *listener=nullptr)
-    :NumberDataField(Type::REAL, true, edit_format, display_format, listener),
+    :NumberDataField(Type::INTEGER, true, edit_format, display_format, listener),
      value(_value), min(_min), max(_max), step(_step) {}
 
   void Set(int _value) {
